Initial setup for handling different nodes
This will build for one node in the following priority:
Cuda -> OpenMP -> Serial

Note this is only working up through Tpetra so far.
Tested on docker for OpenMP and Serial.
Tested on a linux GPU machine for Cuda.

There are some in progress issues to resolve for the interface folder.